requirements.txt
opencv-python
opencv-contrib-python
paho-mqtt
numpy
matplotlib
pyserial
coloredlogs